Robustify teardown or setup for a few tests
* tests/snapshot
* tests/system/test_irqbalance.py
* tests/xapi-plugins/plugin_zfs
