+++ tests/xapi-plugins/plugin_zfs/conftest.py
@@ -1,2 +1,23 @@
+import pytest
+
 # Explicitly import package-scoped fixtures (see explanation in pkgfixtures.py)
-from pkgfixtures import sr_disk_wiped
+from pkgfixtures import host_with_saved_yum_state, sr_disk_wiped
+
+@pytest.fixture(scope='package')
+def host_without_zfs(host):
+    assert not host.file_exists('/usr/sbin/zpool'), \
+        "zfs must not be installed on the host at the beginning of the tests"
+
+@pytest.fixture(scope='package')
+def host_with_zfs(host_without_zfs, host_with_saved_yum_state):
+    host = host_with_saved_yum_state
+    host.yum_install(['zfs'])
+    host.ssh(['modprobe', 'zfs'])
+    yield host
+
+@pytest.fixture(scope='package')
+def zpool_vol0(sr_disk_wiped, host_with_zfs):
+    host_with_zfs.ssh(['zpool', 'create', '-f', 'vol0', '/dev/' + sr_disk_wiped])
+    yield
+    # teardown
+    host_with_zfs.ssh(['zpool', 'destroy', 'vol0'])
